[connman] Notify even when default does not change. Contributes to JB#42337
Call notifier even when default service does not change. This fixes an
important issue of DNS servers not being enabled and causing network not
to work (e.g., bad address from DNS). In order to keep the DNS servers
of the default service enabled notifier must be used to inform
dnsproxy.c and others using the notifier.