[connman] Clear VPN error and state if AutoConnect is set on. JB#47868
Changing AutoConnect value to "on" must have the same effect as user
calling "Connect" method via D-Bus. The failure state is ignored in
do_auto_connect() if user called the VPN to connect.

This fixes the inconsistency with AutoConnect value setting, if user
sets AutoConnect to on clear error and set the service state to idle,
if in failure state. This makes it possible to clear the error and
getting VPN to connect immediately after toggling the AutoConnect on.