pptp: Inform VPN agent about auth errors and unset pwd as hidden value
[pptp] Inform VPN agent on auth err, unset pwd as hidden. JB#48237 JB#45584

Send "VpnAgent.AuthFailure" to VPN agent if there are authentication
errors to indicate that previous login has failed and new credentials
are required to be given. Authentication errors are detected by
plugins/vpn.c:vpn_died() via pptp.c:pptp_error_code() which causes the
error to be indicated accordingly.

If a password is unset with vpn_provider_set_string() a property change
signal will be sent (see vpn/vpn-provider.c:set_string()). This should
be avoided because a VPN agent listening for changes may interpret this
as changing the password into empty string.

This would be the case when authentication fails or when disconnecting.
At the next connection attempt the password retrieved via VPN agent
would be an empty string. Added empty string checks to make sure this
never happens and password is requested from VPN agent when it is also
an empty string.