[nat] Add IPv6 preparation and double NAT support. JB#52923
Implement prepare and restore functionality to for IPv6 NAT. It was
required to add IPv6 type nat struct that is not enabled when nat is
reqularly enabled to maintain the data. The support is enabled by:
 - support setting IPv6 forward for all interfaces
 - support setting accept_ra for the transport interface via ipconfig
 - support adding firewall forwarding rules for the both interfaces
 - support setting ndproxy as enabled with neighbour proxy via inet

Restoring is important to be done when IPv6 setup is not needed. In
some cases where the component/plugin using this functionality uses an
external process which takes time to shut down cleanly and during this
time a service may have removed its ipconfig making restore impossible.
Thus, nat must keep the reference to the given IPv6 ipconfig for the
duration the IPv6 nat is required. This makes it safe for the plugin,
for instance, to keep the ptr to the ipconfig as well between nat IPv6
prepare and restore.

Add also support for double NAT. This is done by adding a second
set of NAT addresses with an additional destination address using the
interface as the MASQUERADE rule out interface. When tethering enables
NAT this new rule is enabled as well since they both are for IPv4. If
the double NAT is enabled when NAT is already running similar restart of
just re-adding the iptables nat table rules is done as it previously
was.

It is also required to use different tracking for the IP forward setting
to facilitate support for both IPv4 and IPv6. This is achieved by
setting the forwarding count as a separate value for both IP family
types and to ensure that the IP forward is disabled only when really
needed.

Also make the free'ing of struct connman_nat a bit better. Using the
common free'ing function eliminates the existing memory leaks in case of
errors.