[ofono] Clear IP-address when protocol does not use it. JB#52923
It is imperative to clear the ipaddress and set the ipconfig method to
off when the particular IP-family is not being set by the protocol. In
other words, when IPv6 only is used there should not be IPv4 address
present and the method should be clearly CONNMAN_IPCONFIG_METHOD_OFF.
Similarly for IPv4 only. This causes confusion when trying to check
whether IPv4 or IPv6 is enabled on a network/service or not as when
changing the protocol the address would remain and the method would
indicate that it is still in use. This change fixes that.