[ofono] Clear IP-address when protocol does not use it. JB#52923
It is imperative to clear the ipaddress when the particular IP family
is not being set by the protocol. In other words, when IPv6 only is
used there should not be IPv4 address present. Similarly for IPv4 only.
This causes confusion when trying to check whether IPv4 or IPv6 is
enabled on a network/service or not as when changing the protocol the
address would remain and the method would indicate that it is still
in use. This change fixes that.

It is unsure whether the method needs to be set as well or does it cause
for confusion. Thus, it is left out as commented reminder to decide what
to do with it later on.