[clat] Keep ptr to ipconfig between IPv6 prep and restore. JB#52923
IPv6 ipconfig reference is held by nat when IPv6 is set. It is,
therefore, safe to have the ptr to it between prepare and restore. This
is mainly because shutting down of tayga may last longer than the
disconnecting service keeps the IPv6 ipconfig or holds the ptr to it (so
connman_service_get_ipconfig() may not return anything). This could also
happen when going to flight mode.