[clat] Support VPN over CLAT, fix tracked service setting. Fixes JB#60276
Add a VPN mode to CLAT by injecting the IPv4 address into the tracked
service's ipconfig and ipaddress. This way the use of CLAT within
ConnMan is more transparent. Routing is mostly handled by ConnMan core
with this change and the host routes set for VPN is handled as in case
when VPN connects over service without the need for CLAT. The ipconfig
is being reset to the CLAT address on the tracked service via service.c's
address reset, which enables the ipconfig for connection.c to use for
routing and gateway setting.

When in VPN mode and in case tethering is or was previously enabled
double nat will not be set as it is not required. Thus, tethering is
done solely over VPN in such scenario.

Also when returning from VPN mode and default changes back to transport
the IPv4 address may have been set by CLAT. Check that and continue if
that is the case.

In VPN mode the VPN service is also tracked via the transport which is
the tracked service. VPN is disconnected by the plugin only when CLAT is
enabled and the VPN tracking is being unset (cleared).

Change to use ipconfig.c IP address check. Plugins should not need to do
this extra. Also do not react to ipconfig change set by CLAT as this is
the side effect of injecting IPv4 of CLAT to the tracked service. The
dual index mode allows to separate interfaces within a service and in
this case the ipconfig is checked to have the same index as the CLAT
device and to ignore the change in such case. Ipconfig changes are
checked always when the state is any of the run states.

Tracked service management is also improved by clearing the IP address
when CLAT is stopped. This makes it a smooth transition to a network
where IPv4 support is also enabled. Additionally The default service can
be sometimes NULL because of reasons. It is not necessarily requiring
to stop CLAT. Therefore, if the tracked service is still running and
CLAT is running ignore the NULL default service.

Use the proper nat.h path for including it. The header is now included
in the Makefile.am properly.