[unit] Update CLAT test for VPN support. JB#60276
This adds VPN as well as tethering tests for verifying that when
cellular/wlan/ethernet is used a transport the cases are handled
accordingly. Tethering is also tested with VPN.

The tracking of added routes is improved for the VPN and tethering
needs in the unit test stubs. As a result of injecting CLAT IPv4 into
tracked service the default route setting is disabled by default and is,
thus, disabled in the tests.

All services have IPv4 ipconfig as they are created. This is now
accounted for in the unit test as well. This required separation of the
ipaddress initialization in tests to be more obvious and flexible.

The default service may be NULL in some cases and as the plugin handles
it correctly now so are the relevant tests changed to do so.

The pausing of the prefix query when VPN is enabled is handled as well
by the test.