[ipconfig] Add helper for checking if ipconfig has an IP address. JB#60276
Add helper function to check if given ipconfig has an IP address set and
it the particular IP address is enabled by method. Returns automatically
false if IP address is omitted and when there is an address set the
method is used to determine if the address is being used.