[ofono] Set ipconfig off if address is cleared, clear address if not enabled. JB#60276
When the protocol is either IPv4 or IPv6 also set the method to OFF to
enforce the status of being off on the not enabled IP family.

When the IP family is not being enabled on DUAL mode clear also the
existing address. This has an effect when changing from DUAL to IPv6.