[service] Add function for plugins to reset ipconfig to addr. JB#60276
To allow reset of ipconfig to new or empty address the existing
functionality for reset with D-Bus message is split to allow
re-usability of code. The new function can be used by plugins and it
sets the address similarly to D-Bus message use but accepts
address/netmask/gateway etc. input. The new ipconfig is enabled after a
succesful reset in order to trigger setting of gateway and routes.