[service] Add ipconfig notify for plugins, support change toggle. JB#60276
Add ipconfig notify via service for plugins with a support for
toggling the change notify via PropertyChanged D-Bus signal. Some
plugins might need more control when the setup requires first resetting
the ipconfig of a service into certain state, which may not be a
connected state, thus the change notify would not be sent. Therefore, it
should be more controlled when the plugin reaches itself a connected
state to inform upper layers that the injection of IP family ipconfig
has been completed successfully.

In case the notify has been already done it is assumed that previously
the ipconfig info isn't notified. The plugin wrapper, therefore, only
amends the functionality of __connman_service_ipconfig_indicate_state()
that does not send the notify when called. This amended functionality is
controlled by a given boolean.

Also done some cleanup.