[clat] Add state changer function with ipconfig notify. JB#60276
Make clat to change the ipconfig state according to CLAT internal state.
All of the states cannot be notified because of how ConnMan does combine
them. Association/configuration/disconnect will be dominant and that
would make cellular disconnect/change state as well. That is not
desired. State combiner would require more work in this case. As of now
the state changer is not checking the state changes but might require
work in the future.