[openvpn] Add support for using untrusted_ip when trusted_ip is not set yet. JB#60276
In some cases the gateway IP may be reported as untrusted_ip variable.
Use that if trusted_ip has not been set. This may help overcome some
setup issues.

In OpenVPN specification on untrusted_ip:
Actual IP address of connecting client or peer which has not been authenticated yet.
Sometimes used to nmap the connecting host in a --tls-verify script to ensure it is
firewalled properly.