doc: Add GetTetheringClientDetails to Manager API
[doc] Add GetTetheringClientDetails to Manager API. JB#60565

Add an experimental method to get the tethering client details with IP
address, IP address type/method, technology used and technology version.

Change the documentation on the TetheringClientsChanged signal to state
that only the added or that have IP address change are listed in the
first array. This will make it more easy to differentiate and notify
about new clients that are connected to tethering interface.