setting: Read only keys that exist from conf.d configs.
[setting] Read only keys that exist from conf.d configs. JB#61966

The additional configs may have a small amount of keys, it is more
optimal to read only what are set. Furthermore, this avoids from
resetting values to default which have a default value, like a string
list for blacklisted interfaces, when additional config is read by going
through all the values.